refractor: move prompt methods into seperate service
