+++ diversitytracker.api/Program.cs
@@ -7,6 +7,7 @@ using diversitytracker.api.Repository;
 using diversitytracker.api.Services;
 using Microsoft.AspNetCore.Authentication.JwtBearer;
 using Microsoft.EntityFrameworkCore;
+using Microsoft.Identity.Client;
 
 var builder = WebApplication.CreateBuilder(args);
 
@@ -27,13 +28,14 @@ builder.Services.AddSwaggerGen();
 
 builder.Services.AddDbContext<AppDbContext>(options =>
 {
-    options.UseSqlServer(builder.Configuration.GetConnectionString("ASPNET_SQL"));
+    options.UseSqlServer(builder.Configuration.GetConnectionString("dockerDb"));
 });
 
 builder.Services.AddScoped<IFormsRepository, FormsDataRepository>();
 builder.Services.AddScoped<IQuestionsRepository, QuestionsRepository>();
 builder.Services.AddScoped<IAiInterpretationRepository, AiInterpretationRepository>();
 builder.Services.AddHttpClient<IAiInterpretationService, AiInterpretationService>();
+builder.Services.AddScoped<IPromptService, PromptService>();
 
 builder.Services.AddAutoMapper(typeof(AutomapperConfig));
